requirements.txt
absl_py==2.0.0
caffe2==0.8.1
numpy==1.26.1
onnx==1.14.1
onnxruntime==1.16.1
onnxsim==0.4.35
opencv_python==4.8.1.78
PyYAML==6.0.1
setuptools==68.0.0
tensorflow==2.14.0
tensorflow_addons==0.22.0
tensorflow_probability==0.22.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
